Rajout de ALT
J'ai rajouté ALT standard dans le cas monodirectionnel (dans le cas bidirectionnel -- ça vaut le coup de creuser le multi héritage en c++ ?)
-- Ça marche sur l'instance petite, sur grande instance il y a un tps important de preprocess d'où : 
TODO = implémenter un moyen de sauvegarde / lecture de fichier de préprocess.

+++ Code/Graphe.cpp
@@ -106,7 +106,7 @@ void Graphe::show(){
 }
 
 
-int Graphe::get_randomSommet(){
+long Graphe::get_randomSommet(){
     map<long, Sommet>::iterator it = V.begin();
     advance(it, rand()%V.size());
     return it->first;
